Move BaseRecordJson into Base namespace
