+++ visitz/VisitzApi/Models/Base/BaseRecordJson.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace VisitzApi.Models;
+namespace VisitzApi.Models.Base;
 
 public class BaseRecordJson
 {
