+++ visitz/VisitzApi/Models/Caseload/IncidentJson.cs
@@ -1,4 +1,5 @@
 using System.Text.Json.Serialization;
+using VisitzApi.Models.Base;
 
 namespace VisitzApi.Models.Caseload;
