+++ visitz/VisitzApiTest/Models/BaseRecordJsonTests.cs
@@ -1,6 +1,6 @@
 using System.Text.Json;
 using VisitzApi.Json;
-using VisitzApi.Models;
+using VisitzApi.Models.Base;
 
 namespace VisitzApiTest.Models;
 public partial class BaseRecordJsonTests
